requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
cffi==1.12.3              # via zstandard
chardet==3.0.4            # via requests
dj-database-url==0.5.0
django==2.2
google-api-core==1.14.2   # via google-cloud-core
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3        # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-bigquery==1.18.0
google-cloud-core==1.0.3  # via google-cloud-bigquery, google-cloud-storage
google-cloud-storage==1.19.0
google-resumable-media==0.3.3  # via google-cloud-bigquery, google-cloud-storage
googleapis-common-protos==1.6.0  # via google-api-core
gunicorn==19.9.0
httplib2==0.13.1          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
lxml==4.4.0
lz4==2.1.10
numpy==1.17.1
pandas==0.25.1
protobuf==3.9.1           # via google-api-core, google-cloud-bigquery, googleapis-common-protos
psycopg2==2.8.2
pyarrow==0.14.1
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.6             # via pyasn1-modules, rsa
pycparser==2.19           # via cffi
python-dateutil==2.8.0    # via pandas
pytz==2018.9              # via django, google-api-core, pandas
pyyaml==5.1.2
requests==2.22.0
rsa==4.0                  # via google-auth
scipy==1.3.1
six==1.12.0               # via google-api-core, google-api-python-client, google-auth, google-resumable-media, protobuf, pyarrow, python-dateutil
sqlparse==0.3.0           # via django
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.3           # via requests
whitenoise==4.1.2
zstandard==0.11.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via google-api-core, protobuf
